Integrate track splitting into existing transcoders
Integrate track splitting into existing transcoders. Insert %S into any transcoder which needs to support track splitting, this placeholder gets replaced by the "Split options" preference which defaults to "-ss %o -t %d" as used by ffmpeg. Integrating splitting into transcoders is more efficient by removing the need for an additional transcoder step; it is also necessary to enable track splitting for formats which can not be split without transcoding (e.g. APE).

This change set also removes the special treatment of downsamplining in TranscodingService.java; the separate Downsample command is still used but this is now treated as just another transcoder to be applied when necessary. It would be possible to replace the "Downsample command' option with a reference to one of the existing transcoders (the current Downsample command is identical to the "mp3 audio" transcoder after all) but this is not done yet.

As an aside this change set also includes an ancient change to JaudiotaggerParser.java to keep it from claiming support for the APE format. There have been several attempts to get this fix implemented in the original Airsonic but somehow it never stuck. Jaudiotagger DOES NOT SUPPORT APE, using it anyway causes APE files to be registered with a duration of 0 which makes them unplayable. By removing APE from the list of supported formats the scanner uses FFmpegParser.java for the format which does work. I will submit a separate PR for this change in the hope that it now, finally, after all those years, can get fixed.

Merge relativepaths